move send_to_closest fn to network
