+++ safenode/src/client/api.rs
@@ -21,9 +21,7 @@ use crate::{
     },
 };
 use bls::{PublicKey, SecretKey, Signature};
-use futures::future::select_all;
 use libp2p::PeerId;
-use std::time::Duration;
 use tokio::task::spawn;
 use xor_name::XorName;
 
@@ -180,14 +178,11 @@ impl Client {
         request: DataRequest,
     ) -> Result<Vec<Result<DataResponse>>> {
         info!("Sending {:?} to the closest peers.", request.dst());
-        let closest_peers = self
+        let responses = self
             .network
-            .client_get_closest_peers(*request.dst().name())
+            .client_send_to_closest(&Request::Data(request))
             .await?;
 
-        let responses = self
-            .send_and_get_responses(closest_peers, &Request::Data(request), true)
-            .await;
         // convert `Response` into `DataResponse`
         let responses = responses
             .into_iter()
@@ -196,7 +191,7 @@ impl Client {
                     Response::Data(res) => Ok(res),
                     _ => Err(Error::DataResponseExpected),
                 },
-                Err(err) => Err(err),
+                Err(err) => Err(err.into()),
             })
             .collect();
 
@@ -212,47 +207,5 @@ impl Client {
                 vec![]
             }
         }
-    }
-
-    // Send a `Request` to the provided set of nodes and wait for their responses concurrently.
-    // If `get_all_responses` is true, we wait for the responses from all the nodes. Will return an
-    // error if the request timeouts.
-    // If `get_all_responses` is false, we return the first successful response that we get.
-    async fn send_and_get_responses(
-        &self,
-        nodes: Vec<PeerId>,
-        req: &Request,
-        get_all_responses: bool,
-    ) -> Vec<Result<Response>> {
-        let mut list_of_futures = Vec::new();
-        for node in nodes {
-            let future = Box::pin(tokio::time::timeout(
-                Duration::from_secs(10),
-                self.network.send_request(req.clone(), node),
-            ));
-            list_of_futures.push(future);
-        }
-
-        let mut responses = Vec::new();
-        while !list_of_futures.is_empty() {
-            match select_all(list_of_futures).await {
-                (Ok(res), _, remaining_futures) => {
-                    let res = res.map_err(Error::Network);
-                    info!("Got response for the req: {req:?}, res: {res:?}");
-                    // return the first successful response
-                    if !get_all_responses && res.is_ok() {
-                        return vec![res];
-                    }
-                    responses.push(res);
-                    list_of_futures = remaining_futures;
-                }
-                (Err(timeout_err), _, remaining_futures) => {
-                    responses.push(Err(Error::ResponseTimeout(timeout_err)));
-                    list_of_futures = remaining_futures;
-                }
-            }
-        }
-
-        responses
     }
 }
